feat: ajustes y nuevos cambios

+++ acciones/v_estudiantes_delete.php
@@ -1,5 +1,10 @@
 <?php
 include("../conexion.php");
 $id     = $_POST['id'];
-$sql    = $con->query("DELETE FROM estudiantes WHERE id = '$id'");
+$sql_count_trabajos = $con->query("SELECT id FROM trabajos WHERE estudiante_id = '$id'");
+if(mysqli_num_rows($sql_count_trabajos) < 1){
+    $sql = $con->query("DELETE FROM estudiantes WHERE id = '$id'");
+}else{
+    $sql = $con->query("UPDATE estudiantes SET estatus = 0 WHERE id = '$id'");
+}
 ?>
